A great day.
Got system TXing and RXing packets. Wrote some great test vector code, so we can directly diff input and output. 

Big problem: at the moment, RX_output will, after a certain packet or two, essentially skip a word and accidentally set its BP too high. This causes it to read in the value after the BP as the length, and all hell breaks loose. 

This could be due to a few things. First, since I'm using the output of the TX_OUT section (GMII vectors are compatable that way), I occasionally get packets separated by a _single cycle_ of TX_EN being 0. Now, I can't find much in the spec, but I'm pretty sure this is legal. 


Therefor, the current thought is maybe RX_INPUT isn't handling these sorts of quick-frames well. Now, it does need an extra few cycles following an ENDF to write the BP at whatnot (4, maybe 5? I should check) but, at the same time, there's the FIFO, which should take up the slack. If it takes 5 cycles to write the BP, etc. (i.e. in overhead) then we should be able to send quite a few of these no-gap frames before the fifo overflows, and in any case we should know about it as an error flag. 

So, I'm working on the test waveform for RX_INPUT to try and make sure the behavior is what i expect. 

Barring that, it could also (surprise surprise) be a problem in RX_OUTPUT/lenr, as is common recently. 




git-svn-id: file:///home/jonas/soma/svn-restore.20080710/hardware/network@313 5dccc5d9-beb7-0310-bf71-cc18d937a3f4