Wow, so, we're really really close to having things work.
The testsuite can send packets, we can read and talk to the phy (don't forget that PHY reg 0 needs to = 0x1140 for things to work!), the ram test passes, everything is chipper. 

But, alas, once we constrained all the pins, things no longer seem to want to synthesize. Worse, it may have been _better_ to not use a clock enable for the control interface. We'll see. 

Current attempt is to try and free up some registers so that P&R has a bit more flexibility. 



git-svn-id: file:///home/jonas/soma/svn-restore.20080710/hardware/network@384 5dccc5d9-beb7-0310-bf71-cc18d937a3f4